Add wolfTPM2_UnsetAuth to clear past authorization sessions when not needed
A way to clear auth sessions that are no longer needed was missing.

For example, after a TPM2_NVRead which uses two auth session slots, there
was no way for the user to clear the second slot afterwards. The only
possibility before was to overwrite with invalid values.

This way a TPM2_NVUndefineSpace would fail right after TPM2_NVRead, just
because there was an auth session not prevously cleared. This can be
handled either by the user manually or the user can use wolfTPM2 wrappers
that handle it now automatically using wolfTPM2_UnsetAuth.

Signed-off-by: Dimitar Tomov <dimi@wolfssl.com>